code improvements and bug fixes
