+++ src/frontend/context/LandingContext.js
@@ -27,7 +27,6 @@ function LandingProvider({ children }) {
 
   const handleSearchSubmit = (e) => {
     e.preventDefault();
-    console.log('search', search);
     if (search.trim().length > 0) {
       dispatch({ type: 'SET_FILTER', payload: 'All' });
       navigate({
